requirements.txt
Django==5.1.3
Markdown==3.7
PyJWT==2.10.0
asgiref==3.8.1
colorama==0.4.6
distlib==0.3.9
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
et_xmlfile==2.0.0
filelock==3.16.1
iniconfig==2.0.0
load-dotenv==0.1.0
mysqlclient==2.2.6
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pymongo==4.10.1
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
tzdata==2024.2
virtualenv==20.27.1
django-environ==0.11.2
cloudinary==1.41.0